+++ src/OpenTelemetry.AutoInstrumentation/Configurations/ConfigurationKeys.cs
@@ -61,12 +61,12 @@ internal static class ConfigurationKeys
     public const string InstrumentationEnabled = "OTEL_DOTNET_AUTO_INSTRUMENTATION_ENABLED";
 
     /// <summary>
-    /// Configuration key for disabling all trace instrumentations.
+    /// Configuration key for enabling all resource detectors.
     /// </summary>
     public const string ResourceDetectorEnabled = "OTEL_DOTNET_AUTO_RESOURCE_DETECTOR_ENABLED";
 
     /// <summary>
-    /// Configuration key template for disabled trace instrumentations.
+    /// Configuration key template for enabling resource detectors.
     /// </summary>
     public const string EnabledResourceDetectorTemplate = "OTEL_DOTNET_AUTO_{0}_RESOURCE_DETECTOR_ENABLED";